+++ src/main/java/de/ukbonn/mwtek/dashboard/controller/DataRetrievalController.java
@@ -102,7 +102,7 @@ public class DataRetrievalController {
 
   public static final String WORKFLOW_ABORTED = "Workflow aborted: ";
   public static final String CURRENT_DATASET_VERSION = "0.5.7";
-  public static final String CURRENT_DDP_VERSION = "0.5.7";
+  public static final String CURRENT_DDP_VERSION = "0.5.7+update.0";
   public static final String FILE_GENERATOR = "ddp";
   public static final String EXPORT_TIMESTAMP = "exporttimestamp";
   public static final String DDP_VERSION = "ddp_version";